backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
zxcvbn==4.4.28
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
supabase==1.0.4

